Lift up the states
